database

+++ database/migrations/2021_03_30_091337_create_users_table.php
@@ -25,7 +25,7 @@ class CreateUsersTable extends Migration
             $table->string('picture_url')->nullable();
             $table->integer('status')->default(1)->comment="0 for deleted, 1 for active";
             $table->integer('verified')->default(0)->comment="0 for unverified, 1 for verified";
-            $table->integer('verify_token');
+            $table->bigInteger('verify_token');
             $table->foreignId('priviledge_id')->nullable();
             $table->timestamp('email_verified_at')->nullable();
             $table->rememberToken();
